+++ packages/moonstone/VideoPlayer/VideoPlayer.less
@@ -391,7 +391,7 @@
 				}
 
 				.sliderFrame {
-					.sliderBar {
+					.bars {
 						background-color: @moon-video-player-slider-bg-color;
 					}
